+++ src/backend/data/mot-data.cpp
@@ -57,7 +57,7 @@ void	motHandler::process_mscGroup (uint8_t	*data,
 	                              bool	lastSegment,
 	                              int16_t	segmentNumber,
 	                              uint16_t	transportId) {
-uint16_t segmentSize	= ((data [0] & 0x1F) << 8) | data [1];
+uint32_t segmentSize	= ((data [0] & 0x1F) << 8) | data [1];
 
 	if ((segmentNumber == 0) && (groupType == 3))  // header
 	   processHeader (transportId, &data [2], segmentSize, lastSegment);
@@ -80,7 +80,7 @@ uint16_t segmentSize	= ((data [0] & 0x1F) << 8) | data [1];
 //	creates a header for an item in a directory
 void	motHandler::processHeader (int16_t	transportId,
 	                           uint8_t	*segment,
-	                           int16_t	segmentSize,
+	                           int32_t	segmentSize,
 	                           bool		lastFlag) {
 uint32_t headerSize	=
 	     ((segment [3] & 0x0F) << 9) | (segment [4]) | (segment [5] >> 7);
@@ -89,7 +89,7 @@ uint32_t bodySize	=
 	                    (segment [2] << 4 ) | ((segment [3] & 0xF0) >> 4);
 uint8_t contentType	= ((segment [5] >> 1) & 0x3F);
 uint16_t contentsubType = ((segment [5] & 0x01) << 8) | segment [6];
-int16_t	pointer	= 7;
+int32_t	pointer	= 7;
 QString	name 	= QString ("");
 
 //	If we had a header with that transportId, do not do anything
@@ -143,7 +143,7 @@ QString	name 	= QString ("");
 void	motHandler::processSegment	(int16_t	transportId,
 	                                 uint8_t	*bodySegment,
 	                                 int16_t	segmentNumber,
-	                                 int16_t	segmentSize,
+	                                 int32_t	segmentSize,
 	                                 bool		lastFlag) {
 int16_t	i;
 
@@ -289,7 +289,7 @@ int16_t	i;
 
 void	motHandler::processDirectory (int16_t	transportId,
                                       uint8_t	*segment,
-                                      int16_t	segmentSize,
+                                      int32_t	segmentSize,
                                       bool	lastFlag) {
 uint32_t directorySize	= ((segment [0] & 0x3F) << 24) |
 	                  ((segment [1]) << 16) |
@@ -318,7 +318,7 @@ int16_t segSize		= ((segment [9] & 0x1F) << 8) |
 void	motHandler::directorySegment (uint16_t	transportId,
                                       uint8_t	*segment,
                                       int16_t	segmentNumber,
-                                      int16_t	segmentSize,
+                                      int32_t	segmentSize,
                                       bool	lastSegment) {
 int16_t	i;
